refactor models.py into models/pydantic_models
