+++ app/core/models/pydantic_models.py
